Split misc section into 2 sections: require node sources and build task. init tests: separated config and personal config tests with multiple describes
